# Conflicts:
#	nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/command/application/holidayshipment/SaveHolidayShipmentCommandHandler.java

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.dom/src/main/java/nts/uk/ctx/at/record/dom/optitem/calculation/CalcResultOfAnyItem.java
@@ -1,5 +1,6 @@
 package nts.uk.ctx.at.record.dom.optitem.calculation;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import lombok.AllArgsConstructor;
@@ -16,9 +17,9 @@ import nts.uk.ctx.at.record.dom.optitem.OptionalItemNo;
 @AllArgsConstructor
 public class CalcResultOfAnyItem {
 	private OptionalItemNo optionalItemNo;
-	Optional<Integer> count;
-	Optional<Integer> time;
-	Optional<Integer> money;
+	Optional<BigDecimal> count;
+	Optional<BigDecimal> time;
+	Optional<BigDecimal> money;
 	
 	
 	/**
@@ -27,7 +28,7 @@ public class CalcResultOfAnyItem {
 	 * @param optionalItemAtr
 	 * @return
 	 */
-	public CalcResultOfAnyItem reCreateCalcResultOfAnyItem(Integer value, OptionalItemAtr optionalItemAtr) {
+	public CalcResultOfAnyItem reCreateCalcResultOfAnyItem(BigDecimal value, OptionalItemAtr optionalItemAtr) {
 		switch (optionalItemAtr) {
 		case NUMBER:
 			return new CalcResultOfAnyItem(this.optionalItemNo,Optional.of(value),this.time,this.money);
